[refactor] 기존 redux 폴더 삭제 및 적용
