--- client/src/Contexts/Bid/constant.js
@@ -1,10 +1,0 @@
-export const PATCH_BID = 'PATCH_BID';
-export const DELETE_BID = 'DELETE_BID';
-
-export const GET_BID_LIST_CUSTOMER = 'GET_BID_LIST_CUSTOMER';
-export const GET_BID_LIST_CUSTOMER_SUCCESS = 'GET_BID_LIST_CUSTOMER_SUCCESS';
-export const GET_BID_LIST_CUSTOMER_ERROR = 'GET_BID_LIST_CUSTOMER_ERROR';
-
-export const GET_BID_LIST_DESIGNER = 'GET_BID_LIST_DESIGNER';
-export const GET_BID_LIST_DESIGNER_SUCCESS = 'GET_BID_LIST_DESIGNER_SUCCESS';
-export const GET_BID_LIST_DESIGNER_ERROR = 'GET_BID_LIST_DESIGNER_ERROR';
