--- client/src/Contexts/Bid/index.js
@@ -1,3 +1,0 @@
-export * from './action';
-export { default } from './reducer';
-export * from './constant';
