--- client/src/Contexts/Branding/constant.js
@@ -1,15 +1,0 @@
-export const GET_BRANDING_LIST_DESIGNER = 'GET_BRANDING_LIST_DESIGNER';
-export const GET_BRANDING_LIST_DESIGNER_SUCCESS = 'GET_BRANDING_LIST_DESIGNER_SUCCESS';
-export const GET_BRANDING_LIST_DESIGNER_ERROR = 'GET_BRANDING_LIST_DESIGNER_ERROR';
-export const GET_MAIN_BRANDING_DESIGNER = 'GET_MAIN_BRANDING_DESIGNER';
-export const GET_MAIN_BRANDING_DESIGNER_SUCCESS = 'GET_MAIN_BRANDING_DESIGNER_SUCCESS';
-export const GET_MAIN_BRANDING_DESIGNER_ERROR = 'GET_MAIN_BRANDING_DESIGNER_ERROR';
-
-export const REGISTER_BRANDING = 'REGISTER_BRANDING';
-export const PATCH_BRANDING = 'PATCH_BRANDING';
-export const PATCH_MAIN_BRANDING = 'PATCH_MAIN_BRANDING';
-export const DELETE_BRANDING = 'DELETE_BRANDING';
-
-export const GET_BRANDING_LIST = 'GET_BRANDING_LIST';
-export const GET_BRANDING_LIST_SUCCESS = 'GET_BRANDING_LIST_SUCCESS';
-export const GET_BRANDING_LIST_ERROR = 'GET_BRANDING_LIST_ERROR';
