--- client/src/Contexts/Branding/index.js
@@ -1,3 +1,0 @@
-export * from './action';
-export { default } from './reducer';
-export * from './constant';
