--- client/src/Contexts/Matching/action.js
@@ -1,21 +1,0 @@
-import {
-  REGISTER_MATCHING,
-  PATCH_MATCHING,
-  DELETE_MATCHING,
-  GET_MATCHING_CUSTOMER,
-  GET_MATCHING_LIST_DESIGNER,
-} from './constant';
-import MatchingAPI from '../../Api/matching';
-import { createPromiseThunk } from '../Common/asyncUtils';
-
-export const registerMatching = (payload) => ({ type: REGISTER_MATCHING, payload });
-export const patchMatching = (id, payload) => ({ type: PATCH_MATCHING, id, payload });
-export const deleteMatching = (id) => ({ type: DELETE_MATCHING, id });
-export const getMatchingByCustomerId = createPromiseThunk(
-  GET_MATCHING_CUSTOMER,
-  MatchingAPI.getMatchingByCustomerId,
-);
-export const getMatchingListByDesignerId = createPromiseThunk(
-  GET_MATCHING_LIST_DESIGNER,
-  MatchingAPI.getMatchingListByDesignerId,
-);
