--- client/src/Contexts/Matching/constant.js
@@ -1,15 +1,0 @@
-export const REGISTER_MATCHING = 'REGISTER_MATCHING';
-export const REGISTER_MATCHING_SUCCESS = 'REGISTER_MATCHING_SUCCESS';
-export const REGISTER_MATCHING_ERROR = 'REGISTER_MATCHING_ERROR';
-export const PATCH_MATCHING = 'PATCH_MATCHING';
-export const PATCH_MATCHING_LIST = 'PATCH_MATCHING_LIST';
-export const DELETE_MATCHING = 'DELETE_MATCHING';
-export const DELETE_MATCHING_LIST = 'DELETE_MATCHING_LIST';
-
-export const GET_MATCHING_CUSTOMER = 'GET_MATCHING_CUSTOMER';
-export const GET_MATCHING_CUSTOMER_SUCCESS = 'GET_MATCHING_CUSTOMER_SUCCESS';
-export const GET_MATCHING_CUSTOMER_ERROR = 'GET_MATCHING_CUSTOMER_ERROR';
-
-export const GET_MATCHING_LIST_DESIGNER = 'GET_MATCHING_LIST_DESIGNER';
-export const GET_MATCHING_LIST_DESIGNER_SUCCESS = 'GET_MATCHING_LIST_DESIGNER_SUCCESS';
-export const GET_MATCHING_LIST_DESIGNER_ERROR = 'GET_MATCHING_LIST_DESIGNER_ERROR';
