--- client/src/Contexts/MatchingHistory/constant.js
@@ -1,13 +1,0 @@
-export const REGISTER_MATCHINGHISTORY = 'REGISTER_MATCHINGHISTORY';
-export const PATCH_MATCHINGHISTORY = 'PATCH_MATCHINGHISTORY';
-export const DELETE_MATCHINGHISTORY = 'DELETE_MATCHINGHISTORY';
-
-export const GET_MATCHINGHISTORY_LIST_CUSTOMER = 'GET_MATCHINGHISTORY_LIST_CUSTOMER';
-export const GET_MATCHINGHISTORY_LIST_CUSTOMER_SUCCESS =
-  'GET_MATCHINGHISTORY_LIST_CUSTOMER_SUCCESS';
-export const GET_MATCHINGHISTORY_LIST_CUSTOMER_ERROR = 'GET_MATCHINGHISTORY_LIST_CUSTOMER_ERROR';
-
-export const GET_MATCHINGHISTORY_LIST_DESIGNER = 'GET_MATCHINGHISTORY_LIST_DESIGNER';
-export const GET_MATCHINGHISTORY_LIST_DESIGNER_SUCCESS =
-  'GET_MATCHINGHISTORY_LIST_DESIGNER_SUCCESS';
-export const GET_MATCHINGHISTORY_LIST_DESIGNER_ERROR = 'GET_MATCHINGHISTORY_LIST_DESIGNER_ERROR';
