--- client/src/Contexts/Proposal/constant.js
@@ -1,17 +1,0 @@
-export const REGISTER_PROPOSAL = 'REGISTER_PROPOSAL';
-export const REGISTER_PROPOSAL_SUCCESS = 'REGISTER_PROPOSAL_SUCCESS';
-export const REGISTER_PROPOSAL_ERROR = 'REGISTER_PROPOSAL_ERROR';
-export const REGISTER_WITHFILE = 'REGISTER_WITHFILE';
-export const REGISTER_WITHFILE_SUCCESS = 'REGISTER_WITHFILE_SUCCESS';
-export const REGISTER_WITHFILE_ERROR = 'REGISTER_WITHFILE_ERROR';
-export const PATCH_PROPOSAL = 'PATCH_PROPOSAL';
-export const DELETE_PROPOSAL = 'DELETE_PROPOSAL';
-export const GET_PROPOSAL = 'GET_PROPOSAL';
-export const GET_PROPOSAL_ASYNC = 'GET_PROPOSAL_ASYNC';
-export const GET_PROPOSAL_ASYNC_SUCCESS = 'GET_PROPOSAL_ASYNC_SUCCESS';
-export const GET_PROPOSAL_ASYNC_ERROR = 'GET_PROPOSAL_ASYNC_ERROR';
-
-// for Proposal List
-export const GET_PROPOSAL_LIST = 'GET_PROPOSAL_LIST';
-export const GET_PROPOSAL_LIST_SUCCESS = 'GET_PROPOSAL_LIST_SUCCESS';
-export const GET_PROPOSAL_LIST_ERROR = 'GET_PROPOSAL_LIST_ERROR';
