--- client/src/Contexts/Style/constant.js
@@ -1,5 +1,0 @@
-export const GET_STYLE_LIST = 'GET_STYLE_LIST';
-export const GET_STYLE_LIST_SUCCESS = 'GET_STYLE_LIST_SUCCESS';
-export const GET_STYLE_LIST_ERROR = 'GET_STYLE_LIST_ERROR';
-
-export const DELETE_STYLE = 'DELETE_STYLE';
