--- client/src/Contexts/StyleScrap/constant.js
@@ -1,6 +1,0 @@
-export const REGISTER_STYLESCRAP = 'REGISTER_STYLESCRAP';
-export const DELETE_STYLESCRAP = 'DELETE_STYLESCRAP';
-
-export const GET_STYLESCRAP_LIST = 'GET_STYLESCRAP_LIST';
-export const GET_STYLESCRAP_LIST_SUCCESS = 'GET_STYLESCRAP_LIST_SUCCESS';
-export const GET_STYLESCRAP_LIST_ERROR = 'GET_STYLESCRAP_LIST_ERROR';
